Flatten Entity class hierarchy
Entity as common base class. No more EntityWithMetadata and
EntityWithSources
