+++ nixio/entity.py
@@ -8,6 +8,9 @@
 # LICENSE file in the root of the Project.
 
 from . import util
+from .section import Section
+from .source import Source
+from .container import LinkContainer
 
 
 class Entity(object):
@@ -18,9 +21,16 @@ class Entity(object):
         self._parent = nixparent
 
     @classmethod
-    def _create_new(cls, nixparent, h5parent):
-        id_ = util.create_id()
-        h5group = h5parent.open_group(id_)
+    def _create_new(cls, nixparent, h5parent, name=None, type_=None):
+        if name and type_:
+            util.check_entity_name_and_type(name, type_)
+            id_ = util.create_id()
+        else:
+            name = util.create_id()
+            id_ = name
+        h5group = h5parent.open_group(name)
+        h5group.set_attr("name", name)
+        h5group.set_attr("type", type_)
         h5group.set_attr("entity_id", id_)
         newentity = cls(nixparent, h5group)
         newentity.force_created_at()
@@ -86,38 +96,21 @@ class Entity(object):
             util.check_attr_type(t, int)
         self._h5group.set_attr("updated_at", util.time_to_str(t))
 
-    def __eq__(self, other):
-        if hasattr(other, "id"):
-            return self.id == other.id
-        else:
-            return False
+    @property
+    def definition(self):
+        """
+        The definition of the entity. The definition can contain a textual
+        description of the entity. This is an optional read-write
+        property, and can be None if no definition is available.
 
-    def __hash__(self):
-        return hash(self.id)
+        :type: str
+        """
+        return self._h5group.get_attr("definition")
 
-
-class NamedEntity(Entity):
-
-    def __init__(self, nixparent, h5group):
-        try:
-            util.check_entity_name_and_type(h5group.get_attr("name"),
-                                            h5group.get_attr("type"))
-            util.check_entity_id(h5group.get_attr("entity_id"))
-        except ValueError:
-            ValueError("Invalid NIX object found in file.")
-        super(NamedEntity, self).__init__(nixparent, h5group)
-
-    @classmethod
-    def _create_new(cls, nixparent, h5parent, name, type_):
-        util.check_entity_name_and_type(name, type_)
-        h5group = h5parent.open_group(name)
-        h5group.set_attr("name", name)
-        h5group.set_attr("type", type_)
-        h5group.set_attr("entity_id", util.create_id())
-        newentity = cls(nixparent, h5group)
-        newentity.force_created_at()
-        newentity.force_updated_at()
-        return newentity
+    @definition.setter
+    def definition(self, d):
+        util.check_attr_type(d, str)
+        self._h5group.set_attr("definition", d)
 
     @property
     def name(self):
@@ -149,20 +142,49 @@ class NamedEntity(Entity):
         self._h5group.set_attr("type", t)
 
     @property
-    def definition(self):
+    def sources(self):
         """
-        The definition of the entity. The definition can contain a textual
-        description of the entity. This is an optional read-write
-        property, and can be None if no definition is available.
-
-        :type: str
+        A property containing all Sources referenced by the group. Sources
+        can be obtained by index or their id. Sources can be removed from the
+        list, but removing a referenced Source will not remove it from the
+        file. New Sources can be added using the append method of the list.
+        This is a read only attribute.
         """
-        return self._h5group.get_attr("definition")
+        if self._sources is None:
+            self._sources = SourceLinkContainer("sources", self, Source,
+                                                self._parent.sources)
+        return self._sources
 
-    @definition.setter
-    def definition(self, d):
-        util.check_attr_type(d, str)
-        self._h5group.set_attr("definition", d)
+    @property
+    def metadata(self):
+        """
+        Associated metadata of the entity. Sections attached to the entity
+        via this attribute can provide additional annotations. This is an
+        optional read-write property, and can be None if no metadata is
+        available.
+
+        :type: Section
+        """
+        if "metadata" in self._h5group:
+            return Section(None, self._h5group.open_group("metadata"))
+        else:
+            return None
+
+    @metadata.setter
+    def metadata(self, sect):
+        if not isinstance(sect, Section):
+            raise TypeError("Error setting metadata to {}. Not a Section."
+                            .format(sect))
+        self._h5group.create_link(sect, "metadata")
+
+    def __eq__(self, other):
+        if hasattr(other, "id"):
+            return self.id == other.id
+        else:
+            return False
+
+    def __hash__(self):
+        return hash(self.id)
 
     def __str__(self):
         return "{}: {{name = {}, type = {}, id = {}}}".format(
@@ -171,3 +193,24 @@ class NamedEntity(Entity):
 
     def __repr__(self):
         return self.__str__()
+
+
+class SourceLinkContainer(LinkContainer):
+
+    # TODO: Sources returned from this container have an incorrect _parent
+    # This is the same issue that we have with Sections. It should probably be
+    # solved the same way
+
+    def append(self, item):
+        if util.is_uuid(item):
+            item = self._inst_item(self._backend.get_by_id(item))
+
+        if not hasattr(item, "id"):
+            raise TypeError("NIX entity or id string required for append")
+
+        if not self._itemstore._parent.find_sources(
+                filtr=lambda x: x.id == item.id
+        ):
+            raise RuntimeError("This item cannot be appended here.")
+
+        self._backend.create_link(item, item.id)
