+++ nixio/section.py
@@ -14,7 +14,7 @@ import functools
 from operator import attrgetter
 from collections import Sequence
 
-from .entity import NamedEntity
+from .entity import Entity
 from .property import Property
 from . import util
 from . import exceptions
@@ -63,7 +63,7 @@ class PropertyProxyList(ProxyList):
                                                 "_delete_property_by_id")
 
 
-class Section(NamedEntity):
+class Section(Entity):
 
     def __init__(self, nixparent, h5group):
         super(Section, self).__init__(nixparent, h5group)
